fix script

+++ scripts/rolling-update.sh
@@ -10,7 +10,9 @@ PREVIOUS_CONTAINER_ID=$("$PNPM" run --silent prod:docker ps --services nest-app 
 
 "$PNPM" prod:docker build nest-app
 
-docker container kill -s SIGTERM "$PREVIOUS_CONTAINER_ID"
-docker container remove -f "$PREVIOUS_CONTAINER_ID"
+if [ -n $PREVIOUS_CONTAINER_ID ]; then
+    docker container kill -s SIGTERM "$PREVIOUS_CONTAINER_ID"
+    docker container remove -f "$PREVIOUS_CONTAINER_ID"
+fi
 
-"$PNPM" prod:docker up -d
+"$PNPM" prod:docker up -d --no-deps nest-app
